% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as}
\alias{as}
\alias{as.data.frame.rcer_raw_metadata}
\alias{as.data.table.rcer_raw_metadata}
\alias{as.data.frame.rcer_raw_record}
\alias{as.data.table.rcer_raw_record}
\alias{as.data.frame.rcer_raw_project}
\alias{as.data.table.rcer_raw_project}
\alias{as.data.frame.rcer_raw_user}
\alias{as.data.table.rcer_raw_user}
\title{as.*}
\usage{
\method{as.data.frame}{rcer_raw_metadata}(x, ...)

\method{as.data.table}{rcer_raw_metadata}(x, ...)

\method{as.data.frame}{rcer_raw_record}(x, ...)

\method{as.data.table}{rcer_raw_record}(x, ...)

\method{as.data.frame}{rcer_raw_project}(x, ...)

\method{as.data.table}{rcer_raw_project}(x, ...)

\method{as.data.frame}{rcer_raw_user}(x, ...)

\method{as.data.table}{rcer_raw_user}(x, ...)
}
\arguments{
\item{x}{any \R object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
Coerce REDCapExporter objects to data.frame or data.tables
}
\details{
These functions are S3 methods for \code{\link{as.data.frame}} and
\code{\link[data.table]{as.data.table}}.
}
