% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{build_col_types}
\alias{build_col_types}
\title{Build Column Types}
\usage{
build_col_types(metadata)
}
\arguments{
\item{metadata}{Project metadata as a \code{data.table}.}
}
\description{
Use the project metadata to build the storage mode for the columns of the
records.
}
\examples{
\dontrun{

library(secret)
uri   <- "https://redcap.ucdenver.edu/api/"
token <-  get_secret("project10734", key = "~/.ssh/vaults")
metadata_raw <- export_content(uri = uri, token = token, content = "metadata")
metadata <- data.table::fread(metadata_raw)
build_col_types(metadata)

token <-  get_secret("project13219", key = "~/.ssh/vaults")
metadata_raw <- export_content(uri = uri, token = token, content = "metadata")
metadata <- data.table::fread(metadata_raw)
build_col_types(metadata)


}
}
