% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_redcap_project.R
\name{build_r_data_package}
\alias{build_r_data_package}
\title{Build R Data Package}
\usage{
build_r_data_package(x, ...)
}
\arguments{
\item{uri}{The URI for the REDCap API.  This is passed to
\code{\link[RCurl]{postForm}}.}

\item{token}{The API token for a REDCap project.}

\item{path}{Path where the exported project source will be
created/overwritten.}

\item{author_roles}{a list naming specific roles for each user id found in
the user table from an exported project.  By default all users with be
contributers ('ctb').  You will need to define a author/creater.}

\item{verbose}{provide messages to tell the user what is happening}
}
\description{
Build a R Data Package from the core conents of a REDCap Project.
}
\details{
To export the data from a REDCap project you will need to have an API Token.
Remember, the token is the equivalent of a username and password.  As such
you should not list the token in plan text.  Several alternative methods for
passing the token to this method will be provided in examples and vignettes.
We strongly encourage the use of the pacakge secret
\url{https://cran.r-project.org/package=secret} to build vaults to store
tokens locally.

The initial export will consist of four pieces of data, the user data,
metadata, proejct info, and records.
}
\examples{
## Please read the vignette for examples:
## vignette(topic = "export", package = "REDCapExporter")

library(REDCapExporter)
# avs_raw_core <- export_core()
data(avs_raw_core)
tmppth <- tempdir()
build_r_data_package(avs_raw_core, tmppth, author_roles = list(dewittp = c("cre", "aut")))
fs::dir_tree(tmppth)

}
