% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate.R
\name{checkmate}
\alias{checkmate}
\title{CheckMate}
\usage{
checkmate(x)
}
\arguments{
\item{x}{a \code{rcer_metadata} or \code{rcer_raw_metadata} object.}
}
\description{
Based on a visual basic macro, checkmate examines the metadata for a project
and reports back on possible protected health information and other data
quality issues.
}
