% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{example_data}
\alias{example_data}
\alias{avs_raw_project_info}
\alias{avs_raw_metadata}
\alias{avs_raw_user}
\alias{avs_raw_record}
\title{Raw Exports From an Example REDCap Project}
\format{An object of class \code{rcer_raw_project} (inherits from \code{character}) of length 1.}
\usage{
avs_raw_project_info

avs_raw_metadata

avs_raw_user

avs_raw_record
}
\description{
Raw Exports From an Example REDCap Project
}
\keyword{datasets}
