% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_redcap_project.R
\name{export_content}
\alias{export_content}
\title{Export Content}
\usage{
export_content(uri, token, content, format, ...)
}
\arguments{
\item{uri}{The URI for the REDCap API.}

\item{token}{The API token for the projedct you want to export from.}

\item{content}{The element to export, see Details.}

\item{format}{The format to return.}

\item{...}{additional arguments passed to \code{\link[RCurl]{postForm}}.}
}
\description{
Export specific data elemetns from REDCap
}
\details{
The \code{content} and \code{format} arguements are used to control the
specific items to be exported, and in what format.  **Review the API
documentation**
}
