% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_redcap_project.R
\name{export_redcap_project}
\alias{export_redcap_project}
\title{Export REDCap Project}
\usage{
export_redcap_project(uri, token, other_exports, path = NULL,
  author_roles = NULL, verbose = TRUE)
}
\arguments{
\item{uri}{The URI for the REDCap API.  This is passed to
\code{\link[RCurl]{postForm}}.}

\item{token}{The API token for a REDCap project.}

\item{other_exports}{Other things to export from REDCap **To be implimented**}

\item{path}{Path where the exported project source will be
created/overwritten.}

\item{verbose}{provide messages to tell the user what is happening}
}
\description{
Using the REDCap API, export parts of the project
}
\details{
To export the data from a REDCap project you will need to have an API Token.
Remember, the token is the equivalent of a username and password.  As such
you should not list the token in plan text.  Several alternative methods for
passing the token to this method will be provided in examples and vignettes.
We strongly encourage the use of the pacakge secret
\url{https://cran.r-project.org/package=secret} to build vaults to store
tokens locally.

The initial export will consist of four pieces of data, the user data,
metadata, proejct info, and records.
}
\examples{
## Please read the vignette for examples:
## vignette(topic = "export", package = "REDCapExporteR")
}
