% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{format_record}
\alias{format_record}
\title{Format Records}
\usage{
format_record(metadata, record, ...)
}
\arguments{
\item{metadata}{a \code{rcer_metadata} or \code{rcer_raw_metadata} object.}

\item{record}{a \code{rcer_raw_record} object.}

\item{...}{other arguments passed to \code{\link{col_type}}}
}
\value{
A \code{data.frame}
}
\description{
Use REDCap project metadata to build a well formatted \code{data.frame} for
the record.
}
