% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCapExporter-package.R
\docType{package}
\name{REDCapExporter-package}
\alias{REDCapExporter}
\alias{REDCapExporter-package}
\title{REDCapExporter}
\description{
Automated Construction of R Data Packages from REDCap Projects
}
\details{
Use the REDCap (Research Electronic Data Capture) API to access (meta)data
from a project and format for immediate use or as part of the source code for
a R Data package.

Please read the vignette for examples and use cases.

\code{vignette(topic = "export", package = "REDCapExporter")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dewittpe/REDCapExporter}
  \item \url{http://www.peteredewitt.com/REDCapExporter/}
}

}
\author{
\strong{Maintainer}: Peter DeWitt \email{peter.dewitt@cuanschutz.edu} (\href{https://orcid.org/0000-0002-6391-0795}{ORCID})

}
\keyword{internal}
