% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCapExporter-package.R
\docType{package}
\name{REDCapExporter}
\alias{REDCapExporter}
\title{REDCapExporter}
\description{
Automated Construction of R Data Packages from REDCap Projects
}
\details{
Use the REDCap (Research Electronic Data Capture) API to access (meta)data
from a project and format for immediate use or as part of the source code for
a R Data package.

Please read the vignette for examples and use cases.

\code{vignette(topic = "export", package = "REDCapExporter")}
}
