% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring.R
\name{REDCapExporter_keyring}
\alias{REDCapExporter_keyring}
\alias{REDCapExporter_keyring_check}
\alias{add_api_token}
\alias{REDCapExporter_get_api_token}
\title{Set up and use of a Keyring for REDCap API Tokens}
\usage{
REDCapExporter_keyring_check(keyring = "REDCapExporter", password = NULL)

add_api_token(
  project,
  keyring = "REDCapExporter",
  user = NULL,
  password = NULL,
  overwrite = FALSE
)

REDCapExporter_get_api_token(
  project,
  keyring = "REDCapExporter",
  user = NULL,
  password = NULL
)
}
\arguments{
\item{keyring}{a character vector identifying the name of the keyring,
defaults to \code{"REDCapExporter"}}

\item{password}{This is the password for the keyring.  The default is an
empty password.}

\item{project}{the name of the REDCap project the API token is identified by.}

\item{user}{user name to associate the token with.  Defaults to
\code{Sys.info()[["user"]]}.}

\item{overwrite}{logical, if \code{TRUE} overwrite the existing token.}
}
\description{
Tools for checking for, and setting up, a file based keyring for storing
REDCap API tokens.
}
