% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_descritption_file}
\alias{write_descritption_file}
\title{Write the DESCRIPTION file for an Exported REDCap Project}
\usage{
write_descritption_file(access_time, user, roles, project_info, path)
}
\description{
Write the DESCRIPTION file for an Exported REDCap Project
}
