% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col-type.R
\name{col_type}
\alias{col_type}
\title{Column Types}
\usage{
col_type(x, factors = TRUE, ...)
}
\arguments{
\item{x}{a \code{rcer_metadata} or \code{rcer_raw_metadata} object}

\item{factors}{If \code{TRUE} (default) then variables reported via drop-down
lists and radio buttons are set up to be \code{factor}s.  If \code{FALSE},
then the column type will be \code{character}.}

\item{...}{not currently used}
}
\value{
a \code{rcer_col_type} object
}
\description{
Define a type for each column of the records for a REDCap project based on
the metadata for the project.
}
\examples{

data("avs_raw_metadata")
col_type(avs_raw_metadata)

}
