% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_record.R
\name{format_record}
\alias{format_record}
\title{Format Record}
\usage{
format_record(record, metadata = NULL, col_type = NULL, ...)
}
\arguments{
\item{record}{a \code{rcer_raw_record} object.}

\item{metadata}{a \code{rcer_metadata} or \code{rcer_raw_metadata} object.
Will be ignored if \code{col_type} is defined.}

\item{col_type}{a \code{rcer_col_type} object.}

\item{...}{other arguments passed to \code{\link{col_type}}}
}
\value{
A \code{data.frame}
}
\description{
Use REDCap project metadata to build a well formatted \code{data.frame} for
the record.
}
\examples{

data("avs_raw_metadata")
data("avs_raw_record")

avs <- format_record(avs_raw_record, avs_raw_metadata)

str(avs)

}
