% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{read_text}
\alias{read_text}
\title{Read Text}
\usage{
read_text(x, class = "data.frame")
}
\arguments{
\item{x}{the raw return from the API call to REDCap}

\item{class}{\code{"data.frame"} (default) or \code{"data.table"}}
}
\value{
a \code{data.frame}
}
\description{
Read raw REDCap API return.  Built to parse csv or json.
}
\details{
This is a non-exported function and not expected to be called by the end
user.  Used by the \code{as.data.frame} methods.
}
